Consolidate game rules into the eval-cell function.
Writing a test for the known Game of Life automata, 'blinker', exposed
the fact that the implementation of the rules was incorrect.

I was not applying all the rules to check for cell death.  So, I wound
up with more cells being alive than I expected.

This test also exposed that testing on a 4x3 grid is not robust enough
to catch an error like this. Especially in a "Wrapped" world where the
edges of the board touch.

Consolidating the game rules into one function makes them easier to
read, and reason about.  I was having a tough time reasoning about the
algorithm in my previous implementation.